% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrpackage.R
\docType{package}
\name{jrAdvPackage-package}
\alias{jrAdvPackage-package}
\alias{jrAdvPackage}
\title{R courses by Jumping Rivers}
\description{
Functions and datasets used in for R
courses run by Jumping Rivers. This package is used in
the "Advanced Programming / Building a package" course".
}
\keyword{package}
